requirements.txt
pyresparser
python-docx
pdfminer.six
find_job_titles
numpy